head_swap/pyproject.toml
[tool.poetry]
name = "face_generation"
version = "0.1.0"
description = ""
authors = ["mo"]

[tool.poetry.dependencies]
python = "^3.7.4"
torch = "^1.3"
torchvision = "^0.4.1"
pytorch-lightning = "^0.5.2"
visdom = "^0.1.8"
dominate = "^2.4"
pillow = "^6.2"
click = "^7.0"
opencv-python = "^4.1"
dlib = "^19.18"
facenet-pytorch = "^0.5.0"
scikit-image = "^0.16.2"
gitpython = "^3.0.7"
keras = "^2.3.1"
tensorflow-estimator = "==2.1.0"
tensorflow = "^2.1.0"
tensorboard = "^2.1.1"
torchgeometry = "^0.1.2"
kornia = "^0.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
